refactor: LottoService의 역할 WinningLotto와 LottoFactory 생성 후 위임
