+++ src/main/kotlin/lotto/service/LottoService.kt
@@ -2,28 +2,23 @@ package lotto.service
 
 import Lottos
 import lotto.domain.Lotto
+import lotto.domain.LottoFactory
 import lotto.domain.LottoNumber
 import lotto.domain.LottoPrice
-import lotto.domain.LottoRandomGenerator
-import lotto.domain.Rank
+import lotto.domain.WinningLotto
 import lotto.domain.WinningResult
 
 class LottoService {
     fun purchase(price: LottoPrice): Lottos {
-        val purchaseCount = price.calculatePurchaseCount()
-        val tickets = (1..purchaseCount).map { LottoRandomGenerator.randomGenerate() }
-        return Lottos(tickets)
+        return LottoFactory.create(price)
     }
 
     fun checkWinning(
         lottos: Lottos,
         winningNumbers: Lotto,
-        bonusBall: LottoNumber,
+        bonusNumber: LottoNumber,
     ): WinningResult {
-        val winningStatistics =
-            lottos.matchNumber(winningNumbers, bonusBall)
-                .filterKeys { it != Rank.NONE }
-
-        return WinningResult(winningStatistics)
+        val winningLotto = WinningLotto(winningNumbers, bonusNumber)
+        return winningLotto.checkWinning(lottos)
     }
 }
